requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.6.2
bs4==0.0.1
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
datasets==2.12.0
dill==0.3.6
dnspython==2.3.0
filelock==3.12.0
Flask==2.3.2
Flask-Cors==3.0.10
frozenlist==1.3.3
fsspec==2023.4.0
graphqlclient==0.2.4
huggingface-hub==0.17.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
loguru==0.7.0
MarkupSafe==2.1.2
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.14
networkx==3.1
numpy==1.24.3
openai==0.27.5
packaging==23.1
pandas==2.0.1
Pillow==10.0.1
pinecone-client==2.2.2
pyarrow==11.0.0
PyAudio==0.2.13
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
regex==2023.8.8
requests==2.29.0
responses==0.18.0
safetensors==0.3.3
six==1.16.0
soupsieve==2.5
sympy==1.12
tokenizers==0.13.3
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
tqdm==4.65.0
transformers==4.33.2
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
Werkzeug==2.3.4
xxhash==3.2.0
yarl==1.9.2
